Partiklene blir nå hentet fra system i stedet for å bli sent rundt (i WaveFunction). Dette er ikke gjort i randomsuniform og distanceCheck + calcualteInterparticleDistances fordi da er ikke partiklene hentet inn i system ennå. Jeg har sjekket at ting funker med og uten optimization og rettet på noen feilmeldinger som dukket opp. Det gav ingen stor speed-up å bruke optimization. Det er noe galt med interaction-delen som gjør at energien av og til blir veldig negativ. Det er forskjellige steder for ulik seed i rng. Et annet problem er at jeg får veldig stor varians når jeg regner den ut på vanlig måte, men mye mindre med blocking-metoden. Jeg skjønner ikke helt hva det kommer av. Jeg har plottet histogramet og den vanlige variansen virker riktig basert på det. Kan ta det med i rapporten. Jeg skal prøve å løse de første oppgavene nå, og så ser jeg på interaction-greiene i morgen.

+++ particle.cpp
@@ -6,7 +6,7 @@ Particle::Particle() {
 }
 
 void Particle::setPosition(const std::vector<double> &position) {
-    assert(position.size() == m_numberOfDimensions);
+    assert((int) position.size() == m_numberOfDimensions);
     m_position = position;
 }
